Made a lot of little adjustments to allow for the chromatin data to work with the config, the config previously loads the gene expression data as g, then I override to deconvolve the chromatin with my own deconvolve function.
May need to refactor this later, but I essentially just wanted to skip the step to remove the 110 timepoint because that timepoint is available in the chromatin data. I didn't think the wt1 data was needed in a bunch of places in the config and model file. But it seems to work okay if we allow it to set ot None. The calculate H just needs to know how many timepoints there are.