Potential fix for DG1 enrichment issue.
It appears that the original smoothing constraints may have not have been proportionally smoothing the top and daughter branches. Smoothing initial and top branch together and daughter separately gave the daughter branch a bit more freedom to absorb extra reads. The current fix is to split the smoothing over initial, top, and bottom branches separately. Upon making this change, and ensuring the smoothing is done properly over the start of DG1 and CG1 from PostG1 and mirroring properly, the initial results look much better. The reads are evenly distributed across the entirety of the F time course.